Visions of perfection normalization
